+++ GraphWebApi/Startup.cs
@@ -17,7 +17,6 @@ using FileService.Interfaces;
 using FileService.Services;
 using GraphExplorerSamplesService.Interfaces;
 using GraphExplorerSamplesService.Services;
-using GraphExplorerPermissionsService.Services;
 using Microsoft.ApplicationInsights.Extensibility.PerfCounterCollector.QuickPulse;
 using Serilog;
 using System;
@@ -60,15 +59,11 @@ namespace GraphWebApi
                 c.DefaultRequestHeaders.Add("User-Agent", "HttpClientFactory-Sample");
             });
 
-
             services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
             services.AddSingleton<ISnippetsGenerator, SnippetsGenerator>();
             services.AddSingleton<IFileUtility, AzureBlobStorageUtility>();
-            services.AddScoped<IPermissionsTestStore, PermissionsTestStore>();
-            services.AddScoped<ISamplesTestStore, SamplesTestStore>();
             services.AddSingleton<IPermissionsStore, PermissionsStore>();
             services.AddSingleton<ISamplesStore, SamplesStore>();
-            //services.AddScoped<ISamplesTestStore, SamplesTestStore>();
             services.Configure<SamplesAdministrators>(Configuration);
 
             #region AppInsights
