+++ GraphWebApi/appsettings.json
@@ -14,13 +14,11 @@
     "InstrumentationKey": "ENTER_KEY_HERE",
     "AppInsightsApiKey": "ENTER_KEY_HERE"
   },
-
   "AllowedHosts": "*",
   "GraphMetadata": {
     "V1.0": "https://raw.githubusercontent.com/microsoftgraph/msgraph-metadata/master/clean_v10_metadata/cleanMetadataWithDescriptionsv1.0.xml",
     "Beta": "https://raw.githubusercontent.com/microsoftgraph/msgraph-metadata/master/clean_beta_metadata/cleanMetadataWithDescriptionsbeta.xml"
-  },  
-
+  },
   "BlobStorage": {
     "ConnectionString": "ENTER_AZURE_STORAGE_CONNECTION_STRING",
     "GithubHost": "https://raw.githubusercontent.com/",
